+++ examples/dialogs/multiline-inputbox.py
@@ -7,7 +7,9 @@ from quo.dialog import InputBox
 
 result = InputBox(
         title="PromptBox shenanigans", 
-        text="What Country are you from?:",)
+        text="What Country are you from?:",
+        multiline=True,
+        bg=False)
 
 print(f"Result = {result}")