+++ src/quo/style/transformation.py
@@ -13,7 +13,7 @@ from abc import ABCMeta, abstractmethod
 from colorsys import hls_to_rgb, rgb_to_hls
 from typing import Callable, Hashable, Optional, Sequence, Tuple, Union
 
-from quo.cache import memoized
+from quo.cache.core import memoized
 from quo.filters import FilterOrBool, to_filter
 from quo.utils.utils import AnyFloat, to_float, to_str