+++ src/quo/cache/__init__.py
@@ -1,2 +1,2 @@
-from .core import SimpleCache, FastDictCache, memoized
-from ._lru_cache import LRUCache
+#from .core import SimpleCache, FastDictCache, memoized
+#from ._lru_cache import LRUCache