+++ src/quo/layout/screen.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 from typing import TYPE_CHECKING, Callable, DefaultDict, Dict, List, NamedTuple, Optional, Tuple
 
-from quo.cache import FastDictCache
+from quo.cache.core import FastDictCache
 from quo.utils.utils import get_width
 
 if TYPE_CHECKING: