+++ src/quo/style/style.py
@@ -7,7 +7,7 @@ import sys
 from enum import Enum
 from typing import Dict, Hashable, List, Set, Tuple, TypeVar
 
-from quo.cache import SimpleCache
+from quo.cache.core import SimpleCache
 
 from .core import (
     ANSI_COLOR_NAMES,